requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-core
langchain-openai
fastmcp
python-dotenv
pyyaml
requests
feedparser
yfinance
pydantic
beautifulsoup4
